Take certain liberties with the recommended base16 palette
* Use base05 (default foreground, ANSI color 7) for ColorTable[07]
  instead of base07 (white)
* Use base03 (comments, ANSI color 8) for ColorTable[08] instead of
  base01 (lighter background)
* Use base07 (white, ANSI color 15) for ColorTable[15] instead of
  base06 (deprecated text)

These changes re-align ColorTable colors to ANSI palette and the base16
colors that are traditionally used in this palette.

In addition, colors in ColorTable[09:14] have been replaced with their
"darker" counterparts from ColorTable[01:06]. This achieves the effect of
disabling the use of bright colors for bold text in ANSI-based terminals.
For example, this prevents Git for Windows from using colors in the range of
ColorTable[09:14] for bolded text, which causes bright red to be displayed
in orange and bright green to be displayed in base02 (selection
background).